+++ app/Http/Kernel.php
@@ -68,5 +68,6 @@ class Kernel extends HttpKernel
         'isAdmin' => \App\Http\Middleware\isAdmin::class,
         'isTechniker' => \App\Http\Middleware\isTechniker::class,
         'isCustomer' => \App\Http\Middleware\isCustomer::class,
+        'agent' => \App\Http\Middleware\AuthenticateAgent::class,
     ];
 }